+++ disassemblers/ofrak_pyghidra/CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased 0.2.0rc6](https://github.com/redballoonsecurity/ofrak/tree/master)
 
 ### Added
-- Support `ProgramAttributes` `entry_points` and `base_address` fields for passing program metadata to PyGhidra
-- Support `MemoryRegionPermissions` attribute for fine-grained memory region permission control
+- Support `ProgramAttributes` `entry_points` and `base_address` fields for passing program metadata to PyGhidra ([#701](https://github.com/redballoonsecurity/ofrak/pull/701))
+- Support `MemoryRegionPermissions` attribute for fine-grained memory region permission control ([#701](https://github.com/redballoonsecurity/ofrak/pull/701))
 - Add a PyGhidra custom load analyzer to allow for loading programs with a custom layout ([#677](https://github.com/redballoonsecurity/ofrak/pull/677))
 - Add detailed logging output and progress indicators to standalone analysis script ([#672](https://github.com/redballoonsecurity/ofrak/pull/672))